Configured IDP for secure API and return access token with Identity Claims. Configured Client to send JWT Bearer Token on each requests when request to API.

+++ ImageGallery.API/Controllers/ImagesController.cs
@@ -31,8 +31,14 @@ namespace ImageGallery.API.Controllers
         [HttpGet()]
         public async Task<ActionResult<IEnumerable<Image>>> GetImages()
         {
+            var ownerId = User.Claims.FirstOrDefault(p => p.Type == "sub")?.Value;
+            if (ownerId == null)
+            {
+                throw new Exception("User identifier is missing from token.");
+            }
+
             // get from repo
-            var imagesFromRepo = await _galleryRepository.GetImagesAsync();
+            var imagesFromRepo = await _galleryRepository.GetImagesAsync(ownerId);
 
             // map to model
             var imagesToReturn = _mapper.Map<IEnumerable<Image>>(imagesFromRepo);
@@ -43,9 +49,14 @@ namespace ImageGallery.API.Controllers
 
         [HttpGet("{id}", Name = "GetImage")]
         public async Task<ActionResult<Image>> GetImage(Guid id)
-        {          
-            var imageFromRepo = await _galleryRepository.GetImageAsync(id);
+        {
+            var ownerId = User.Claims.FirstOrDefault(p => p.Type == "sub")?.Value;
+            if (ownerId == null)
+            {
+                throw new Exception("User identifier is missing from token.");
+            }
 
+            var imageFromRepo = await _galleryRepository.GetImageAsync(id, ownerId);
             if (imageFromRepo == null)
             {
                 return NotFound();
@@ -57,6 +68,7 @@ namespace ImageGallery.API.Controllers
         }
 
         [HttpPost()]
+        [Authorize(Roles = "PayingUser")]
         public async Task<ActionResult<Image>> CreateImage([FromBody] ImageForCreation imageForCreation)
         {
             // Automapper maps only the Title in our configuration
@@ -84,6 +96,12 @@ namespace ImageGallery.API.Controllers
             // ownerId should be set - can't save image in starter solution, will
             // be fixed during the course
             //imageEntity.OwnerId = ...;
+            var ownerId = User.Claims.FirstOrDefault(p => p.Type == "sub")?.Value;
+            if (ownerId == null)
+            {
+                throw new Exception("User identifier is missing from token.");
+            }
+            imageEntity.OwnerId= ownerId;
 
             // add and save.  
             _galleryRepository.AddImage(imageEntity);
@@ -99,16 +117,21 @@ namespace ImageGallery.API.Controllers
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteImage(Guid id)
-        {            
-            var imageFromRepo = await _galleryRepository.GetImageAsync(id);
+        {
+            var ownerId = User.Claims.FirstOrDefault(p => p.Type == "sub")?.Value;
+            if (ownerId == null)
+            {
+                throw new Exception("User identifier is missing from token.");
+            }
 
+            var imageFromRepo = await _galleryRepository.GetImageAsync(id, ownerId);
             if (imageFromRepo == null)
             {
                 return NotFound();
             }
 
             _galleryRepository.DeleteImage(imageFromRepo);
-
+            
             await _galleryRepository.SaveChangesAsync();
 
             return NoContent();
@@ -118,7 +141,13 @@ namespace ImageGallery.API.Controllers
         public async Task<IActionResult> UpdateImage(Guid id, 
             [FromBody] ImageForUpdate imageForUpdate)
         {
-            var imageFromRepo = await _galleryRepository.GetImageAsync(id);
+            var ownerId = User.Claims.FirstOrDefault(p => p.Type == "sub")?.Value;
+            if (ownerId == null)
+            {
+                throw new Exception("User identifier is missing from token.");
+            }
+
+            var imageFromRepo = await _galleryRepository.GetImageAsync(id, ownerId);
             if (imageFromRepo == null)
             {
                 return NotFound();
